Reorganizing the project tree
- Improved the code organization, creating the 'common' and 'extrato' folders
- Updated the templates
